+++ x11-dl/src/lib.rs
@@ -37,6 +37,7 @@ pub mod xmd;
 pub mod xmu;
 pub mod xrecord;
 pub mod xrender;
+pub mod xshm;
 pub mod xss;
 pub mod xt;
 pub mod xtest;